requirements.txt
Django
django-secure
django-registration
django-ipware
django-widget-tweaks
django-dbbackup
mysqlclient
psycopg2
Celery
numpy
redis
pandas
matplotlib
scipy
pyyaml
